+++ pkg/searchquery/searchquery.go
@@ -31,6 +31,7 @@ const (
 	// Temporary experimental fields:
 	FieldIndex   = "index"
 	FieldCount   = "count" // Searches that specify `count:` will fetch at least that number of results, or the full result set
+	FieldMax     = "max"   // Deprecated alias for count
 	FieldTimeout = "timeout"
 )
 
@@ -60,6 +61,7 @@ var (
 			// Experimental fields:
 			FieldIndex:   {Literal: types.StringType, Quoted: types.StringType, Singular: true},
 			FieldCount:   {Literal: types.StringType, Quoted: types.StringType, Singular: true},
+			FieldMax:     {Literal: types.StringType, Quoted: types.StringType, Singular: true},
 			FieldTimeout: {Literal: types.StringType, Quoted: types.StringType, Singular: true},
 		},
 		FieldAliases: map[string]string{